requirements.txt
Django==2.2
django-mdeditor==0.1.13
Markdown==3.1
Pillow==6.0.0
Pygments==2.3.1
PyMySQL==0.9.3
pytz==2019.1
sqlparse==0.3.0
